h3: change priority type and add optional parser
The `quiche::h3::send_response_with_priority()` method allows
applications to specify a stream priority that is used by quiche's
stream prioritization.

Previously, the priority was passed as a String value that was expected
to contain Extensible Priority parameters in a serialized Structured
Fields Dictionary format; see
https://datatracker.ietf.org/doc/html/draft-ietf-httpbis-priority-12#section-4.
Internally, the String was manually parsed to extract the "u" and "i"
parameters.

With this change, we cconvert the priority parameter from a String to a
new `quiche::h3::Priority` type that holds only the parameters we use
inside quiche. Applications are now required to handle Extensible
Priorities field values themselves. A new utility method
`quiche::h3::parse_extensible_priority` has been added to support
application parsing. This relies on the sfv crate, which can be included
as an optional dependency for quiche (and thus controlled via features).
sfv offers a Structured Fields compliant parser, if it fails to read the
expected Extensible Priority format, an error is returned. If the format
is ok but values are outside the allowed range, they get clamped. Using
sfv already caught a few bugs in the old approach and I've added test
cases for those.